[BACK, UPDATE] UserService.java - user가 삭제된 경우 map 삭제

+++ back/src/main/java/com/DormitoryBack/domain/member/domain/service/UserService.java
@@ -249,9 +249,11 @@ public class UserService {
         else if(groupService.isBelongToAnywhere(usrId)){
             throw new RuntimeException("CannotDeleteUserWhileGrouping");
         }
-        
+
         userRepository.delete(target);
-        encryptedEmailService.deleteEmailMap(target.getEncryptedEmail());
-        encryptedPhoneNumService.deleteNumberMap(target.getEncryptedPhoneNum());
+        if(userRepository.findById(usrId)==null){
+            encryptedEmailService.deleteEmailMap(target.getEncryptedEmail());
+            encryptedPhoneNumService.deleteNumberMap(target.getEncryptedPhoneNum());
+        }
     }
 }
